+++ packages/appcd/src/cli.js
@@ -8,7 +8,7 @@ const version = getAppcdVersion();
 let banner;
 if (!Object.prototype.hasOwnProperty.call(process.env, 'APPC_NPM_VERSION')) {
 	banner = `${appcdLogger.styles.highlight('Appcelerator Daemon')}, version ${version}\n`
-		+ 'Copyright (c) 2015-2019, Axway, Inc. All Rights Reserved.';
+		+ 'Copyright (c) 2015-2020, Axway, Inc. All Rights Reserved.';
 }
 
 export default new CLI({
@@ -19,8 +19,8 @@ export default new CLI({
 	helpExitCode: 2,
 	name: 'appcd',
 	options: {
-		'--config <json>':      { type: 'json', desc: 'serialized JSON string to mix into the appcd config' },
-		'--config-file <file>': { type: 'file', desc: 'path to a appcd JS config file' }
+		'--config [json]':      { type: 'json', desc: 'serialized JSON string to mix into the appcd config' },
+		'--config-file [file]': { type: 'file', desc: 'path to a appcd JS config file' }
 	},
 	version
 });